refactor ClcaSamplerErrorTracker AlphaMart uses SamplerTracker eliminate most uses of Sampler
