+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/betting/AlphaMart.kt
@@ -154,16 +154,17 @@ class TruncShrinkage(
 class AlphaMart(
     val estimFn : EstimFn,  // estimator of the population mean
     val N: Int,             // max number of cards for this contest (diluted)
+    val tracker: SamplerTracker,
     val withoutReplacement: Boolean = true,
     val riskLimit: Double = 0.05, // α ∈ (0, 1)
     val upperBound: Double = 1.0,  // aka u
 ): RiskMeasuringFn {
-    val betting: BettingMartOld
+    val betting: BettingMart2
 
     init {
-        val tracker = ClcaErrorTracker(0.0, upperBound) // TODO using ClcaErrorTracker, why not PluralityErrorTracker?
+        // val tracker = ClcaErrorTracker(0.0, upperBound) // TODO using ClcaErrorTracker, why not PluralityErrorTracker?
         val bettingFn = EstimAdapter(N, withoutReplacement, upperBound, estimFn)
-        betting = BettingMartOld(bettingFn, N,  tracker, upperBound, riskLimit)
+        betting = BettingMart2(bettingFn, N,  tracker, upperBound, riskLimit)
     }
 
     override fun testH0(
