+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/betting/SamplerTracker.kt
@@ -3,9 +3,7 @@ package org.cryptobiotic.rlauxe.betting
 import io.github.oshai.kotlinlogging.KotlinLogging
 import org.cryptobiotic.rlauxe.audit.AuditableCard
 import org.cryptobiotic.rlauxe.core.*
-import org.cryptobiotic.rlauxe.estimate.CardSamples
 import org.cryptobiotic.rlauxe.util.Welford
-import org.cryptobiotic.rlauxe.util.df
 import org.cryptobiotic.rlauxe.util.doubleIsClose
 import org.cryptobiotic.rlauxe.util.doublePrecision
 import kotlin.random.Random
@@ -39,13 +37,13 @@ interface SampleErrorTracker: SampleTracker {
 //// For polling audits. Production runPollingAuditRound
 class PollingSamplerTracker(
     val contestId: Int,
-    val cvrPairs: List<Pair<CvrIF, CvrIF>>, // Pair(mvr, card)
     val assorter: AssorterIF,
-    val allowReset: Boolean = true,
-    maxSampleIndexIn: Int? = null,
+    val cvrPairs: List<Pair<CvrIF, CvrIF>>, // Pair(mvr, card)
+    maxSampleIndex: Int? = null,
 ): SamplerTracker {
-    val maxSampleIndex = maxSampleIndexIn ?: cvrPairs.size
+    val useMaxSampleIndex = maxSampleIndex ?: cvrPairs.size
     val permutedIndex = MutableList(cvrPairs.size) { it }
+    val allowReset: Boolean = maxSampleIndex == null
 
     var maxSamples: Int = 0
     private var idx = 0
@@ -54,11 +52,11 @@ class PollingSamplerTracker(
     init {
         cvrPairs.forEach { (mvr, card) ->
             require(mvr.location() == card.location())  { "mvr location ${mvr.location()} != card.location ${card.location()}"}  }
-        maxSamples = cvrPairs.take(maxSampleIndex).count { it.second.hasContest(contestId) }
+        maxSamples = cvrPairs.take(useMaxSampleIndex).count { it.second.hasContest(contestId) }
     }
 
     override fun sample(): Double {
-        while (idx < cvrPairs.size && idx < maxSampleIndex) {
+        while (idx < cvrPairs.size && hasNext()) {
             val (mvr, card) = cvrPairs[permutedIndex[idx]]
             idx++
             if (card.hasContest(contestId)) {
@@ -67,18 +65,18 @@ class PollingSamplerTracker(
                 return lastVal!!
             }
         }
-        logger.error{"PollingSampling no samples left for contest ${contestId} and Assorter ${assorter}"}
-        throw RuntimeException("PollingSampling no samples left for contest ${contestId} and assorter ${assorter}")
+        logger.error{"PollingSamplerTracker no samples left for contest ${contestId} and Assorter ${assorter.shortName()}"}
+        throw RuntimeException("PollingSamplerTracker no samples left for contest ${contestId} and assorter ${assorter.shortName()}")
     }
 
     override fun reset() {
         if (!allowReset) {
-            logger.error {"PollingSampling reset not allowed; contest ${contestId} assorter ${assorter}\""}
-            throw RuntimeException("PollingSampling reset not allowed")
+            logger.error {"PollingSamplerTracker reset not allowed; contest ${contestId} assorter ${assorter.shortName()}\""}
+            throw RuntimeException("PollingSamplerTracker reset not allowed")
         }
         permutedIndex.shuffle(Random)
         idx = 0
-        maxSamples = cvrPairs.take(maxSampleIndex).count { it.second.hasContest(contestId) }
+        maxSamples = cvrPairs.take(useMaxSampleIndex).count { it.second.hasContest(contestId) }
         welford = Welford()
     }
 
@@ -93,79 +91,6 @@ class PollingSamplerTracker(
     var lastVal: Double? = null
     override fun numberOfSamples() = welford.count
     override fun welford() = welford
-    override fun done() {
-        if (lastVal != null) welford.update(lastVal!!)
-        lastVal = null
-    }
-
-    ///////////////////////////////// temporary
-    override fun sum() = welford.sum()
-    override fun mean() = welford.mean
-    override fun variance() = welford.variance()
-
-    override fun last(): Double = lastVal!!
-    override fun addSample(sample: Double) {
-        TODO("Not implemented")
-    }
-}
-
-class ClcaSamplerTracker(
-    val contestId: Int,
-    val cvrPairs: List<Pair<CvrIF, AuditableCard>>, // Pair(mvr, card)
-    val cassorter: ClcaAssorter,
-    val allowReset: Boolean,
-    maxSampleIndexIn: Int? = null,
-): SamplerTracker {
-    val maxSampleIndex = maxSampleIndexIn ?: cvrPairs.size
-    val permutedIndex = MutableList(cvrPairs.size) { it }
-
-    private var maxSamples = 0
-    private var idx = 0
-    private var welford = Welford()
-
-    init {
-        cvrPairs.forEach { (mvr, card) ->
-            require(mvr.location() == card.location())  { "mvr location ${mvr.location()} != card.location ${card.location()}"}  }
-        maxSamples = cvrPairs.take(maxSampleIndex).count { it.second.hasContest(contestId) }
-    }
-
-    override fun sample(): Double {
-        while (idx < cvrPairs.size && idx < maxSampleIndex) {
-            val (mvr, card) = cvrPairs[permutedIndex[idx]]
-            idx++
-            if (card.hasContest(contestId)) {
-                if (lastVal != null) welford.update(lastVal!!)
-                lastVal = cassorter.bassort(mvr, card, hasStyle=card.exactContests())
-                return lastVal!!
-            }
-        }
-        logger.error{"ClcaSampling no samples left for ${contestId} and ComparisonAssorter ${cassorter}"}
-        throw RuntimeException("ClcaSampling no samples left for ${contestId} and ComparisonAssorter ${cassorter}")
-    }
-
-    override fun reset() {
-        if (!allowReset) {
-            logger.error{"ClcaSampling reset not allowed ; contest ${contestId} cassorter ${cassorter}"}
-            throw RuntimeException("ClcaSampling reset not allowed")
-        }
-        permutedIndex.shuffle(Random)
-        idx = 0
-        maxSamples = cvrPairs.take(maxSampleIndex).count { it.second.hasContest(contestId) }
-        welford = Welford()
-    }
-
-    override fun maxSamples() = maxSamples
-    override fun maxSampleIndexUsed() = idx
-    override fun nmvrs() = cvrPairs.size
-
-    override fun hasNext() = (welford.count + 1 < maxSamples)
-    override fun next() = sample()
-
-    // tracker reflects "previous sequence"
-    var lastVal: Double? = null
-    override fun numberOfSamples() = welford.count
-    override fun welford() = welford
-
     override fun done() {
         if (lastVal != null) welford.update(lastVal!!)
         lastVal = null
@@ -185,149 +110,42 @@ class ClcaSamplerTracker(
 //// For clca audits. Production RunClcaContestTask
 class ClcaSamplerErrorTracker(
     val contestId: Int,
-    val cvrPairs: List<Pair<CvrIF, AuditableCard>>, // Pair(mvr, card)
     val cassorter: ClcaAssorter,
-    val allowReset: Boolean,
-    maxSampleIndexIn: Int? = null,  // if maxSampleIndexIn != null then reset = false
+    val samples: List<Pair<CvrIF, AuditableCard>>, // Pair(mvr, card)
+    val allowReset: Boolean = true,  // needed ?
 ): SamplerTracker, SampleErrorTracker {
-    val maxSampleIndex = maxSampleIndexIn ?: cvrPairs.size
-    val permutedIndex = MutableList(cvrPairs.size) { it }
-    val clcaErrorTracker = ClcaErrorTracker2(cassorter.noerror, cassorter.assorter.upperBound())
-    // val firstDebug: Pair<Int, Int>
-
-    private var maxSamples = 0
-    private var idx = 0
-    private var welford = Welford()
-
-    init {
-        cvrPairs.forEach { (mvr, card) ->
-            require(mvr.location() == card.location())  { "mvr location ${mvr.location()} != card.location ${card.location()}"}  }
-        maxSamples = cvrPairs.take(maxSampleIndex).count { it.second.hasContest(contestId) }
-        // firstDebug = debug()
-    }
-
-    override fun sample(): Double {
-        while (idx < cvrPairs.size && idx < maxSampleIndex) {
-            val (mvr, card) = cvrPairs[permutedIndex[idx]]
-            idx++
-            if (card.hasContest(contestId)) {
-                val nextVal = cassorter.bassort(mvr, card, hasStyle=card.exactContests())
-                clcaErrorTracker.addSample(nextVal, card.poolId == null) // dont track errors from oa pools
-                if (lastVal != null) welford.update(lastVal!!)
-                lastVal = nextVal
-                return nextVal
-            }
-        }
-        logger.error{"ClcaSampling no samples left for ${contestId} and ComparisonAssorter ${cassorter}"}
-        throw RuntimeException("ClcaSampling no samples left for ${contestId} and ComparisonAssorter ${cassorter}")
-    }
-
-    override fun reset() {
-        if (!allowReset) {
-            logger.error{"ClcaSampling reset not allowed ; contest ${contestId} cassorter ${cassorter}"}
-            throw RuntimeException("ClcaSampling reset not allowed")
-        }
-        permutedIndex.shuffle(Random)
-        maxSamples = cvrPairs.take(maxSampleIndex).count { it.second.hasContest(contestId) }
-        welford = Welford()
-        clcaErrorTracker.reset()
-        idx = 0
-        lastVal = null
-    }
-
-    fun debug(): Pair<Int, Int> {
-        var debugIdx = 0
-        var countHasContest = 0
-        var countNotPooled = 0
-        while (debugIdx < cvrPairs.size && debugIdx < maxSampleIndex) {
-            val (mvr, card) = cvrPairs[permutedIndex[debugIdx]]
-            debugIdx++
-            if (card.hasContest(contestId)) {
-                countHasContest++
-                if (card.poolId == null) countNotPooled++
-            }
-        }
-        println("$contestId ${cassorter.shortName()}: countHasContest=$countHasContest countNotPooled=$countNotPooled " +
-                "pct=${df(countNotPooled / countHasContest.toDouble())}")
-        return Pair(countHasContest, countNotPooled)
-    }
-
-    override fun maxSamples() = maxSamples
-    override fun maxSampleIndexUsed() = idx
-    override fun nmvrs() = cvrPairs.size
-
-    override fun hasNext() = (welford.count + 1 < maxSamples)
-    override fun next() = sample()
-
-    // tracker reflects "previous sequence"
-    var lastVal: Double? = null
-    override fun numberOfSamples() = welford.count
-    override fun welford() = welford
-
-    override fun done() {
-        if (lastVal != null) welford.update(lastVal!!)
-        lastVal = null
-    }
-
-    override fun measuredClcaErrorCounts(): ClcaErrorCounts = clcaErrorTracker.measuredClcaErrorCounts()
-    override fun noerror(): Double = clcaErrorTracker.noerror
-
-    ///////////////////////////////// temporary TODO remove
-    override fun sum() = welford.sum()
-    override fun mean() = welford.mean
-    override fun variance() = welford.variance()
-
-    override fun last(): Double = lastVal!!
-    override fun addSample(sample: Double) {
-        TODO("Remove this")
-    }
-}
-
-
-//// For clca audits. Production RunClcaContestTask
-class ClcaSamplerErrorTracker2(
-    val contestId: Int,
-    cvrPairs: List<Pair<CvrIF, AuditableCard>>, // Pair(mvr, card)
-    cardSamples: CardSamples,
-    val cassorter: ClcaAssorter,
-    val allowReset: Boolean,
-): SamplerTracker, SampleErrorTracker {
-    val samples = cardSamples.extractSubsetByIndex(contestId, cvrPairs)
     val permutedIndex = MutableList(samples.size) { it }
     val clcaErrorTracker = ClcaErrorTracker2(cassorter.noerror, cassorter.assorter.upperBound())
-    val firstDebug: Pair<Int, Int>
 
     private var idx = 0
     private var welford = Welford()
 
     init {
         samples.forEach { (mvr, card) ->
-            require(mvr.location() == card.location())  { "mvr location ${mvr.location()} != card.location ${card.location()}"}  }
-        firstDebug = debug()
+            require(mvr.location() == card.location())  { "mvr location ${mvr.location()} != card.location ${card.location()}"}
+            require(card.hasContest(contestId))  { " card.location ${card.location()} does not have contest $contestId" }
+        }
     }
 
     override fun sample(): Double {
         while (idx < samples.size) {
             val (mvr, card) = samples[permutedIndex[idx]]
             idx++
-            if (card.hasContest(contestId)) {
-                val nextVal = cassorter.bassort(mvr, card, hasStyle=card.exactContests())
-                clcaErrorTracker.addSample(nextVal, card.poolId == null) // dont track errors from oa pools
-                if (lastVal != null) welford.update(lastVal!!)
-                lastVal = nextVal
-                return nextVal
-            }  else {
-                logger.error{"cardSamples for contest ${contestId} list card not contining the contest at index ${permutedIndex[idx-1]}"}
-            }
+            val nextVal = cassorter.bassort(mvr, card, hasStyle=card.exactContests())
+
+            clcaErrorTracker.addSample(nextVal, card.poolId == null) // dont track errors from oa pools
+            if (lastVal != null) welford.update(lastVal!!)
+            lastVal = nextVal
+            return nextVal
         }
-        logger.error{"ClcaSampling no samples left for ${contestId} and ComparisonAssorter ${cassorter}"}
-        throw RuntimeException("ClcaSampling no samples left for ${contestId} and ComparisonAssorter ${cassorter}")
+        logger.error{"ClcaSamplerErrorTracker no samples left for contest ${contestId} and ComparisonAssorter ${cassorter.shortName()}"}
+        throw RuntimeException("ClcaSamplerErrorTracker no samples left for ${contestId} and ComparisonAssorter ${cassorter.shortName()}")
     }
 
     override fun reset() {
         if (!allowReset) {
-            logger.error{"ClcaSampling reset not allowed ; contest ${contestId} cassorter ${cassorter}"}
-            throw RuntimeException("ClcaSampling reset not allowed")
+            logger.error{"ClcaSamplerErrorTracker reset not allowed ; contest ${contestId} cassorter ${cassorter.shortName()}"}
+            throw RuntimeException("ClcaSamplerErrorTracker reset not allowed")
         }
         permutedIndex.shuffle(Random)
         welford = Welford()
@@ -336,28 +154,11 @@ class ClcaSamplerErrorTracker2(
         lastVal = null
     }
 
-    fun debug(): Pair<Int, Int> {
-        var debugIdx = 0
-        var countHasContest = 0
-        var countNotPooled = 0
-        while (debugIdx < samples.size) {
-            val (mvr, card) = samples[permutedIndex[debugIdx]]
-            debugIdx++
-            if (card.hasContest(contestId)) {
-                countHasContest++
-                if (card.poolId == null) countNotPooled++
-            }
-        }
-        println("$contestId ${cassorter.shortName()}: countHasContest=$countHasContest countNotPooled=$countNotPooled " +
-                "pct=${df(countNotPooled / countHasContest.toDouble())}")
-        return Pair(countHasContest, countNotPooled)
-    }
-
     override fun maxSamples() = samples.size
-    override fun maxSampleIndexUsed() = idx
+    override fun maxSampleIndexUsed() = idx  // TODO no longer the index in the entire set ...
     override fun nmvrs() = samples.size
 
-    override fun hasNext() = (welford.count + 1 < samples.size)
+    override fun hasNext() = (idx < samples.size)
     override fun next() = sample()
 
     // tracker reflects "previous sequence"
@@ -382,77 +183,56 @@ class ClcaSamplerErrorTracker2(
     override fun addSample(sample: Double) {
         TODO("Remove this")
     }
-}
-
-//// For clca audits with styles and no errors
-class ClcaNoErrorSamplerTracker(
-    val contestId: Int,
-    val contestNc: Int,
-    val cassorter: ClcaAssorter,
-    val cvrIterator: Iterator<CvrIF>,
-): SamplerTracker {
-    private var idx = 0
-    private var done = false
-    private val welford = Welford()
-
-    override fun sample(): Double {
-        while (cvrIterator.hasNext()) {
-            val cvr = cvrIterator.next()
-            idx++
-            if (cvr.hasContest(contestId)) {
-                val result = cassorter.bassort(cvr, cvr, hasStyle=true)
-                welford.update(result)
-                return result
-            }
-        }
-        done = true
-        if (!warned) {
-            logger.warn { "ClcaNoErrorIterator no samples left for ${contestId} and ComparisonAssorter ${cassorter}" }
-            warned = true
-        }
-        return 0.0
-    }
-
-    override fun reset() {
-        logger.error{"ClcaNoErrorIterator reset not allowed"}
-        throw RuntimeException("ClcaNoErrorIterator reset not allowed")
-    }
-
-    override fun numberOfSamples() = welford.count
-    override fun welford() = welford
-
-    override fun done() {}
-
-    override fun maxSamples() = contestNc
-    override fun maxSampleIndexUsed() = idx
-    override fun nmvrs() = contestNc
-
-    override fun hasNext() = !done && cvrIterator.hasNext()
-    override fun next() = sample()
+    ///////////////////////////////
 
     companion object {
-        var warned = false
-    }
+        fun fromIndexList(
+            contestId: Int,
+            cassorter: ClcaAssorter,
+            pairs: List<Pair<AuditableCard, AuditableCard>>,
+            wantIndices: List<Int>,
+        ): ClcaSamplerErrorTracker {
+            val extract = mutableListOf<Pair<AuditableCard, AuditableCard>>()
+            var wantIdx = 0
+            pairs.forEachIndexed { idx, pair ->
+                if (wantIdx < wantIndices.size && idx == wantIndices[wantIdx]) {
+                    extract.add(pair)
+                    wantIdx++
+                }
+            }
+            return ClcaSamplerErrorTracker(contestId, cassorter, extract)
+        }
 
-    ///////////////////////////////// temporary
-    override fun last(): Double {
-        TODO("Not yet implemented")
-    }
+        fun withMaxSample(
+            contestId: Int,
+            cassorter: ClcaAssorter,
+            cvrPairs: List<Pair<CvrIF, AuditableCard>>, // Pair(mvr, card)
+            maxSampleIndex: Int? = null,  // if maxSampleIndexIn != null then reset = false)
+        ): ClcaSamplerErrorTracker {
+            val maxIndex = maxSampleIndex ?: Int.MAX_VALUE
+            val extract = mutableListOf<Pair<CvrIF, AuditableCard>>()
+            cvrPairs.forEachIndexed { idx, pair ->
+                if (pair.second.hasContest(contestId) && idx < maxIndex) {
+                    extract.add(pair)
+                }
+            }
+            return ClcaSamplerErrorTracker(contestId, cassorter, extract)
+        }
 
-    override fun sum(): Double {
-        TODO("Not yet implemented")
-    }
-
-    override fun mean(): Double {
-        TODO("Not yet implemented")
-    }
-
-    override fun variance(): Double {
-        TODO("Not yet implemented")
-    }
-
-    override fun addSample(sample: Double) {
-        TODO("Not yet implemented")
+        fun withNoErrors(
+            contestId: Int,
+            cassorter: ClcaAssorter,
+            cvrIterator: Iterator<AuditableCard>,
+        ): ClcaSamplerErrorTracker {
+            val extract = mutableListOf<Pair<CvrIF, AuditableCard>>()
+            while (cvrIterator.hasNext()) {
+                val cvr = cvrIterator.next()
+                if (cvr.hasContest(contestId)) {
+                    extract.add(Pair(cvr, cvr))
+                }
+            }
+            return ClcaSamplerErrorTracker(contestId, cassorter, extract)
+        }
     }
 }
