+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/estimate/PollingFuzzSampler.kt
@@ -18,6 +18,8 @@ import kotlin.collections.component2
 import kotlin.collections.forEach
 import kotlin.random.Random
 
+// could try to create a subclass of PollingSamplerTracker ??
+
 class PollingFuzzSamplerTracker(
     val fuzzPct: Double,
     val cvrs: List<Cvr>,
