+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/workflow/PersistedWorkflow.kt
@@ -93,7 +93,7 @@ class PersistedWorkflow(
             AuditType.CLCA -> runClcaAuditRound(config, auditRound, mvrManager, auditRound.roundIdx, auditor = ClcaAssertionAuditor(quiet))
             AuditType.POLLING -> runPollingAuditRound(config, auditRound, mvrManager, auditRound.roundIdx, quiet)
             AuditType.ONEAUDIT -> runClcaAuditRound(config, auditRound, mvrManager, auditRound.roundIdx,
-                auditor = OneAuditAssertionAuditor2(mvrManager().oapools()!!, quiet))
+                auditor = OneAuditAssertionAuditor(mvrManager().oapools()!!, quiet))
         }
 
         auditRound.auditWasDone = true
