+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/betting/TestAlphaMart.kt
@@ -4,12 +4,9 @@ import org.cryptobiotic.rlauxe.util.doublePrecision
 import org.cryptobiotic.rlauxe.util.mean2margin
 import org.cryptobiotic.rlauxe.audit.AuditConfig
 import org.cryptobiotic.rlauxe.audit.AuditType
-import org.cryptobiotic.rlauxe.core.AssorterIF
 import org.cryptobiotic.rlauxe.core.ContestWithAssertions
-import org.cryptobiotic.rlauxe.core.CvrIF
 import org.cryptobiotic.rlauxe.estimate.MultiContestTestData
 import org.cryptobiotic.rlauxe.estimate.runRepeatedAlphaMart
-import org.cryptobiotic.rlauxe.workflow.PollingSampler
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
@@ -40,7 +37,7 @@ class TestAlphaMart {
 
         val cvrs = test.makeCvrsFromContests()
         val config = AuditConfig(AuditType.POLLING, nsimEst=10)
-        val samplerTracker = PollingSamplerTracker(contestUA.contest.id, cvrs.zip(cvrs), assorter)
+        val samplerTracker = PollingSamplerTracker(contestUA.contest.id, assorter, cvrs.zip(cvrs))
 
         val eta0 = assorter.dilutedMean()
         println("eta0=$eta0, margin=${mean2margin(eta0)}")
