+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/estimate/TestConsistentSampler.kt
@@ -90,51 +90,4 @@ class TestConsistentSampler {
             assertTrue(contest.estMvrs <= ballotsForContest)
         }
     }
-
-    /*
-    @Test
-    fun testUniformPollingSampling() {
-        val N = 20000
-        val test = MultiContestTestDataP(20, 11, N)
-        val contestsUAs: List<ContestUnderAudit> = test.contests.map { ContestUnderAudit(it, isClca = false).addStandardAssertions() }
-        val contestRounds = contestsUAs.map{ contest -> ContestRound(contest, 1) }
-        contestRounds.forEach { it.estSampleSize = 100 + Random.nextInt(it.Nc/2) }
-
-        val cvrs = test.makeCvrsFromContests()
-        val mvrManager = MvrManagerForTesting(cvrs, cvrs, Random.nextLong())
-
-        val contestSampleCutoff = 10000
-        val config = AuditConfig(AuditType.CLCA, true, contestSampleCutoff = contestSampleCutoff)
-
-        val auditRound = AuditRound(1, contestRounds, samplePrns = emptyList())
-        uniformSampling(auditRound, mvrManager, previousSamples=emptySet(), config, 0)
-        println("nsamples needed = ${auditRound.samplePrns.size}\n")
-
-        // must be ordered
-        var lastRN = 0L
-        auditRound.samplePrns.forEach {
-            require(it > lastRN)
-            lastRN = it
-        }
-
-        contestRounds.forEach { contest ->
-            println(" ${contest.name} (${contest.id}) estSampleSize=${contest.estSampleSize}")
-        }
-
-        // double check the number of cvrs == sampleSize
-        println("contest.name (id) == sampleSize")
-        contestRounds.forEach { contest ->
-            assertTrue(contest.estSampleSizeEligibleForRemoval() <= auditRound.samplePrns.size)
-
-            // fails here sometimes TODO logic is heonous
-            println("done ${contest.done} or estSampleSize=${contest.estSampleSizeNoStyles} <=? ${auditRound.samplePrns.size}")
-            if (contest.estSampleSizeNoStyles > auditRound.samplePrns.size) {
-                // assertTrue(contest.done)
-            }
-
-            println("contest ${contest.id} estSampleSize=${contest.estSampleSizeEligibleForRemoval()} <=? ${contestSampleCutoff}")
-            if (contest.estSampleSizeEligibleForRemoval() > contestSampleCutoff)
-                assertTrue(contest.done)
-        }
-    } */
 }
