+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/shangrla/TestPopulationMean.kt
@@ -150,10 +150,10 @@ class TestPopulationMeanWithoutReplacement {
         assertEquals(0.2448979591836735, assortValues.variance(), doublePrecision)
 
         // compare directly to ALPHA
-        val estimFn = TruncShrinkage(N = N, upperBound = u, d = d, eta0 = eta)
-        val alpha = AlphaMart(estimFn = estimFn, N = N, upperBound = u)
-
         val sampler = SampleFromArray(x.toDoubleArray())
+        val estimFn = TruncShrinkage(N = N, upperBound = u, d = d, eta0 = eta)
+        val alpha = AlphaMart(estimFn = estimFn, N = N, sampler, upperBound = u)
+
         println("alphaTestH0")
         alpha.testH0(x.size, false, ) { sampler.sample() }
