+++ core/src/testFixtures/kotlin/org/cryptobiotic/rlauxe/workflow/TestSampling.kt
@@ -18,108 +18,6 @@ import kotlin.random.Random
 
 private val logger = KotlinLogging.logger("TestSampling")
 
-//// For clca audits with styles and no errors
-class ClcaNoErrorIterator(
-    val contestId: Int,
-    val contestNc: Int,
-    val cassorter: ClcaAssorter,
-    val cvrIterator: Iterator<CvrIF>,
-): Sampler, Iterator<Double> {
-    private var idx = 0
-    private var count = 0
-    private var done = false
-
-    override fun sample(): Double {
-        while (cvrIterator.hasNext()) {
-            val cvr = cvrIterator.next()
-            idx++
-            if (cvr.hasContest(contestId)) {
-                val result = cassorter.bassort(cvr, cvr, hasStyle=true)
-                count++
-                return result
-            }
-        }
-        done = true
-        if (!warned) {
-            logger.warn { "ClcaNoErrorIterator no samples left for ${contestId} and ComparisonAssorter ${cassorter}" }
-            warned = true
-        }
-        return 0.0
-    }
-
-    override fun reset() {
-        logger.error{"ClcaNoErrorIterator reset not allowed"}
-        throw RuntimeException("ClcaNoErrorIterator reset not allowed")
-    }
-
-    override fun maxSamples() = contestNc
-    override fun maxSampleIndexUsed() = idx
-    override fun nmvrs() = contestNc
-
-    override fun hasNext() = !done && cvrIterator.hasNext()
-    override fun next() = sample()
-
-    companion object {
-        var warned = false
-    }
-}
-
-// what about a function to fuzz the cvr or mvr on the fly ??
-class OneAuditNoErrorIterator(
-    val contestId: Int,
-    val contestNc: Int,
-    val contestSampleCutoff: Int?,
-    val cassorter: ClcaAssorter,
-    cvrIter: Iterator<Cvr>,
-): Sampler, Iterator<Double> {
-    val cvrs = mutableListOf<Cvr>()
-    var permutedIndex = mutableListOf<Int>()
-
-    private var idx = 0
-    private var count = 0
-    private var done = false
-
-    init {
-        while ((contestSampleCutoff == null || cvrs.size < contestSampleCutoff) && cvrIter.hasNext()) {
-            val cvr = cvrIter.next()
-            if (cvr.hasContest(contestId)) cvrs.add(cvr)
-        }
-        permutedIndex = MutableList(cvrs.size) { it }
-    }
-
-    override fun sample(): Double {
-        while (idx < cvrs.size) {
-            val cvr = cvrs[permutedIndex[idx]]
-            idx++
-            val result = cassorter.bassort(cvr, cvr, hasStyle=true)
-            count++
-            return result
-        }
-        if (!warned) {
-            logger.warn { "OneAuditNoErrorIterator no samples left for ${contestId} and ComparisonAssorter ${cassorter}" }
-            warned = true
-        }
-        return 0.0
-    }
-
-    override fun reset() {
-        permutedIndex.shuffle(Random)
-        idx = 0
-        count = 0
-    }
-
-    override fun maxSamples() = contestNc
-    override fun maxSampleIndexUsed() = idx
-    override fun nmvrs() = contestNc
-
-    override fun hasNext() = !done && (count < contestNc)
-    override fun next() = sample()
-
-    companion object {
-        var warned = false
-    }
-}
-
 private val show = true
 
 // TODO assumes plurality assorter
@@ -754,57 +652,3 @@ class ClcaFuzzSampler(
     override fun hasNext(): Boolean = (idx < N)
     override fun next(): Double = sample()
 }
-
-/* TODO no longer used by simulateSampleSizeOneAuditAssorter
-class OneAuditFuzzSampler(
-    val fuzzPct: Double,
-    val cvrs: List<Cvr>,
-    val contestUA: OAContestUnderAudit,
-    val cassorter: ClcaAssorter
-): Sampling, Iterator<Double> {
-    val maxSamples = cvrs.count { it.hasContest(contestUA.id) }
-    val N = cvrs.size
-    val permutedIndex = MutableList(N) { it }
-    val welford = Welford()
-    val stratumNames : Set<String>
-    var cvrPairs: List<Pair<Cvr, Cvr>> // (mvr, cvr)
-    var idx = 0
-
-    init {
-        stratumNames = emptySet() // TODO contestUA.contest.pools.values.map { it.name }.toSet() // TODO
-        val mvrs = remakeFuzzed()
-        cvrPairs = mvrs.zip(cvrs)
-    }
-
-    override fun sample(): Double {
-        while (idx < N) {
-            val (mvr, cvr) = cvrPairs[permutedIndex[idx]]
-            idx++
-            if (cvr.hasContest(contestUA.id)) {
-                val result = cassorter.bassort(mvr, cvr)
-                welford.update(result)
-                return result
-            }
-        }
-        throw RuntimeException("no samples left for ${contestUA.id} and ComparisonAssorter ${cassorter}")
-    }
-
-    override fun reset() {
-        val mvrs = remakeFuzzed()
-        cvrPairs = mvrs.zip(cvrs)
-        permutedIndex.shuffle(Random)
-        idx = 0
-    }
-
-    fun remakeFuzzed(): List<Cvr> {
-        return makeFuzzedCvrsFrom(listOf(contestUA.contest), cvrs, fuzzPct) { !stratumNames.contains(it.id) }
-    }
-
-    override fun maxSamples() = maxSamples
-    override fun maxSampleIndexUsed() = idx
-    override fun nmvrs() = idx // TODO
-
-    override fun hasNext(): Boolean = (idx < N)
-    override fun next(): Double = sample()
-} */
-
