+++ core/src/testFixtures/kotlin/org/cryptobiotic/rlauxe/workflow/WorkflowTesterClca.kt
@@ -5,7 +5,7 @@ import org.cryptobiotic.rlauxe.audit.AuditRound
 import org.cryptobiotic.rlauxe.audit.AuditRoundIF
 import org.cryptobiotic.rlauxe.audit.AuditType
 import org.cryptobiotic.rlauxe.audit.AuditableCard
-import org.cryptobiotic.rlauxe.betting.ClcaSamplerTracker
+import org.cryptobiotic.rlauxe.betting.ClcaSamplerErrorTracker
 import org.cryptobiotic.rlauxe.betting.SamplerTracker
 import org.cryptobiotic.rlauxe.core.ClcaAssorter
 import org.cryptobiotic.rlauxe.core.Contest
@@ -64,5 +64,10 @@ fun makeClcaNoErrorSampler(contestId: Int, cvrs : List<Cvr>, cassorter: ClcaAsso
 fun makeClcaNoErrorSamplerTracker(contestId: Int, cvrs : List<Cvr>, cassorter: ClcaAssorter): SamplerTracker {
     val cards = cvrs.mapIndexed { idx, it -> AuditableCard.fromCvr(it, idx, 0) }
     val cvrPairs = cvrs.zip(cards)
-    return ClcaSamplerTracker(contestId, cvrPairs, cassorter, true)
+    return ClcaSamplerErrorTracker.withMaxSample(
+        contestId,
+        cassorter,
+        cvrPairs,
+        maxSampleIndex = null,
+    )
 }
