--- plots/src/test/kotlin/org/cryptobiotic/rlauxe/alpha/CompareShrinkTrunkWithFixed.kt
@@ -1,145 +1,0 @@
-package org.cryptobiotic.rlauxe.alpha
-
-import org.cryptobiotic.rlauxe.betting.AlphaMart
-import org.cryptobiotic.rlauxe.betting.ClcaErrorTracker
-import org.cryptobiotic.rlauxe.betting.FixedEstimFn
-import org.cryptobiotic.rlauxe.betting.PollingSamplerTracker
-import org.cryptobiotic.rlauxe.betting.SamplerTracker
-import org.cryptobiotic.rlauxe.betting.TestH0Result
-import org.cryptobiotic.rlauxe.betting.TruncShrinkage
-import org.cryptobiotic.rlauxe.core.*
-import org.cryptobiotic.rlauxe.workflow.PollingSampler
-import org.cryptobiotic.rlauxe.workflow.Sampler
-import org.cryptobiotic.rlauxe.estimate.makeCvrsByExactMean
-import org.cryptobiotic.rlauxe.makeStandardPluralityAssorter
-import org.cryptobiotic.rlauxe.estimate.RunRepeatedResult
-import org.cryptobiotic.rlauxe.estimate.runRepeated
-import org.cryptobiotic.rlauxe.util.makeContestsFromCvrs
-import org.cryptobiotic.rlauxe.util.mean2margin
-
-import kotlin.test.Test
-
-// CANDIDATE FOR REMOVAL
-
-class CompareShrinkTrunkWithFixed {
-
-    @Test
-    fun compareShrinkTruncWithFixedExact() {
-        val etas = listOf(.505, .51, .52, .55, .6, .7) // alternative means
-        val N = 10000
-        etas.forEach { eta ->
-            val cvrs = makeCvrsByExactMean(N, eta)
-            val pairs = cvrs.zip(cvrs)
-
-            val contestUA = ContestWithAssertions(makeContestsFromCvrs(cvrs).first()).addStandardAssertions()
-
-            val sampleFn = PollingSampler(contestUA.id, pairs, makeStandardPluralityAssorter(N))
-
-            println("\neta0 = $eta")
-            val fixResult = testAlphaMartFixed(eta, sampleFn)
-            println("fixResult=$fixResult")
-
-            sampleFn.reset()
-            val alphaResult = testAlphaMartTrunc(eta, sampleFn)
-            println("alphaResult=$alphaResult")
-        }
-    }
-
-    fun testAlphaMartTrunc(eta0: Double, sampling: Sampler): TestH0Result {
-        val u = 1.0
-        val d = 10000
-        val N = sampling.maxSamples()
-
-        val trunc = TruncShrinkage(N = N, upperBound = u, d = d, eta0 = eta0)
-        val alpha = AlphaMart(estimFn = trunc, N = N)
-
-        return alpha.testH0(N, true) { sampling.sample() }
-    }
-
-    fun testAlphaMartFixed(eta0: Double, sampling: Sampler): TestH0Result {
-        val fixed = FixedEstimFn(eta0 = eta0)
-        val alpha = AlphaMart(estimFn = fixed, N = sampling.maxSamples())
-        return alpha.testH0(sampling.maxSamples(), true) { sampling.sample() }
-    }
-
-    @Test
-    fun compareShrinkTruncWithFixedRepeated() {
-        val etas = listOf(.505, .51, .52, .53, .54, .55, .6, .7) // alternative means
-        val N = 10000
-        val ntrials = 1000
-        val fixResults = mutableListOf<RunRepeatedResult>()
-        val truncResults = mutableListOf<RunRepeatedResult>()
-
-        etas.forEach { eta ->
-            val cvrs = makeCvrsByExactMean(N, eta)
-            val pairs = cvrs.zip(cvrs)
-
-            val contestUA = ContestWithAssertions(makeContestsFromCvrs(cvrs).first()).addStandardAssertions()
-            val sampleFn = PollingSamplerTracker(contestUA.id, pairs, makeStandardPluralityAssorter(N))
-
-            println("\neta0 = $eta")
-            val fixResult = runAlphaMartFixedRepeated(eta, sampleFn, ntrials)
-            println("fixResult=$fixResult")
-            fixResults.add(fixResult)
-
-            sampleFn.reset()
-            val truncResult = runAlphaMartTruncRepeated(eta, sampleFn, ntrials)
-            println("truncResult=$truncResult")
-            truncResults.add(truncResult)
-        }
-
-        calcMean(" pctSamplesNeeded", fixResults, truncResults) { it.pctSamplesNeeded() }
-        // calcMean(" success10percent", fixResults, truncResults) { it.percentHist!!.cumul(10) }
-        // calcMean(" success20percent", fixResults, truncResults) { it.percentHist!!.cumul(20) }
-        // calcMean(" success30percent", fixResults, truncResults) { it.percentHist!!.cumul(30) }
-        // calcMean(" success40percent", fixResults, truncResults) { it.percentHist!!.cumul(40) }
-        // calcMean(" success50percent", fixResults, truncResults) { it.percentHist!!.cumul(50) }
-    }
-
-    fun calcMean(title: String, fixResults : List<RunRepeatedResult>, truncResults : List<RunRepeatedResult>,
-                 fld: (RunRepeatedResult) -> Double) {
-        val fixFld = mutableListOf<Double>()
-        val truncFld = mutableListOf<Double>()
-        fixResults.forEachIndexed{ idx, fixResult ->
-            fixFld.add(fld(fixResult))
-            truncFld.add(fld(truncResults[idx]))
-        }
-        println("ArithmeticMean for $title: fix=${fixFld.average().toInt()}, trunc=${truncFld.average().toInt()}")
-        // println("GeometricMean for $title: fix=${geometricMean(fixFld)}, trunc=${geometricMean(truncFld)}")
-    }
-
-    fun runAlphaMartTruncRepeated(eta0: Double, samplerTracker: SamplerTracker, ntrials: Int): RunRepeatedResult {
-        val u = 1.0
-        val d = 10000
-        val N = samplerTracker.maxSamples()
-
-        val trunc = TruncShrinkage(N = N, upperBound = u, d = d, eta0 = eta0)
-        val alpha = AlphaMart(estimFn = trunc, N = samplerTracker.maxSamples())
-
-        return runRepeated(
-            name = "runAlphaMartTruncRepeated",
-            terminateOnNullReject = true,
-            ntrials = ntrials,
-            testFn = alpha,
-            testParameters = mapOf("eta0" to eta0, "d" to d.toDouble(), "margin" to mean2margin(eta0)),
-            N=N,
-            samplerTracker=samplerTracker,
-            )
-    }
-
-    fun runAlphaMartFixedRepeated(eta0: Double, samplerTracker: SamplerTracker, ntrials: Int): RunRepeatedResult {
-        val N = samplerTracker.maxSamples()
-        val fixed = FixedEstimFn(eta0 = eta0)
-        val alpha = AlphaMart(estimFn = fixed, N = N)
-
-        return runRepeated(
-            name = "runAlphaMartFixedRepeated",
-            terminateOnNullReject = true,
-            ntrials = ntrials,
-            testFn = alpha,
-            testParameters = mapOf("eta0" to eta0, "margin" to mean2margin(eta0)),
-            N=N,
-            samplerTracker = samplerTracker,
-        )
-    }
-}
