+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/estimate/PlotDistributions.kt
@@ -2,7 +2,6 @@ package org.cryptobiotic.rlauxe.estimate
 
 import org.cryptobiotic.rlauxe.testdataDir
 import org.cryptobiotic.rlauxe.audit.*
-import org.cryptobiotic.rlauxe.betting.ClcaErrorTracker2
 import org.cryptobiotic.rlauxe.betting.ClcaSamplerErrorTracker
 import org.cryptobiotic.rlauxe.core.ClcaAssertion
 import org.cryptobiotic.rlauxe.core.ClcaErrorTable
@@ -128,13 +127,12 @@ class PlotDistributions {
             val cassertion = assertionRound.assertion as ClcaAssertion
             val cassorter = cassertion.cassorter
 
-            val sampler = ClcaSamplerErrorTracker(
+            val sampler = ClcaSamplerErrorTracker.withMaxSample(
                 contestUA.id,
-                sortedPairs,
                 cassorter,
-                allowReset = false,
+                sortedPairs,
+                maxSampleIndex = null,
             )
-            // val sampler = ClcaSampler(contestUA.id, sortedPairs.size, sortedPairs, cassorter, allowReset = false) // OK
 
             val contestRound = ContestRound(contestUA, listOf(assertionRound), 1)
             ClcaAssertionAuditor().run(auditConfig, contestRound, assertionRound, sampler, 1)
