+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/workflow/SFoaSingleRoundAuditTaskGenerator.kt
@@ -5,6 +5,7 @@ import org.cryptobiotic.rlauxe.audit.AuditConfig
 import org.cryptobiotic.rlauxe.audit.AuditableCard
 import org.cryptobiotic.rlauxe.audit.ContestRound
 import org.cryptobiotic.rlauxe.audit.PopulationIF
+import org.cryptobiotic.rlauxe.betting.ClcaSamplerErrorTracker
 import org.cryptobiotic.rlauxe.core.Cvr
 import org.cryptobiotic.rlauxe.betting.TestH0Result
 import org.cryptobiotic.rlauxe.estimate.ConcurrentTaskG
@@ -66,9 +67,8 @@ class SfoaSingleRoundAuditTask(
                 )
 
                 val sampler =
-                    ClcaNoErrorIterator(
+                    ClcaSamplerErrorTracker.withNoErrors(
                         contestUA.id,
-                        contestUA.Nc,
                         cassertion.cassorter,
                         mvrManager.sortedCards().iterator(),
                     )
